Made a new battle folder since we'll have to change the battle format for most of the fights.
