+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
 	"python.analysis.extraPaths": [
-		"./items"
+		"./items",
+		"./characters"
 	]
 }
